[modify] slot data
