--- Assets/Externals/KCW/ScriptableObject/ItemData.cs
@@ -1,45 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public enum ItemType
-{
-    Resource,
-    Equipable,
-    Consumable
-}
-
-public enum ConsumableType
-{
-    Hunger,
-    Health
-}
-
-[System.Serializable]
-public class ItemDataConsumable
-{
-    public ConsumableType type;
-    public float value;
-}
-
-[CreateAssetMenu(fileName = "Item", menuName = "New Item")]
-public class ItemData : ScriptableObject
-{
-    [Header("Info")]
-    public string displayName;
-    public string description;
-    public ItemType type;
-    public Sprite icon;
-    public GameObject dropPrefab;
-
-    [Header("Stacking")]
-    public bool canStack;
-    public int maxStackAmount;
-
-    [Header("Consumable")]
-    public ItemDataConsumable[] consumables;
-
-    [Header("Equip")]
-    public GameObject equipPrefab;
-
-}
