--- Assets/Externals/KCW/ScriptableObject/ItemData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4e6dd7dd25b017b4f9b62105b93b24ba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
