+++ Assets/Resources/Prefabs/StoreData.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b17a37fbc9a65934d80a02c7a2a7a3eb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
