+++ Assets/Resources/Prefabs/UI/ItemSlot 2link.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 673606bbebc42944fa15b9be172d94a4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
