+++ Assets/Scenes/KCW_CreateScene.unity
@@ -574,3 +574,64 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 74268a50407f8c84291a44e35359cb0a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1001 &1524735429465565671
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_RootOrder
+      value: 7
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 666
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 760.9999
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 19038990149938986, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3594776430121174499, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
+      propertyPath: m_Name
+      value: StoreData
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: b17a37fbc9a65934d80a02c7a2a7a3eb, type: 3}
