+++ Assets/ScriptableObjects/ItemData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c4ef00a41fbf580419e12aff47d7213e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
