+++ Assets/ScriptableObjects/ItemData/ItemSO.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4b3188d34f0a56b419e342dd4a54006b
+guid: 2279cffc04ed1214ab4750319117313c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
