+++ Assets/ScriptableObjects/ItemData/item_atk10.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d2b776adc655e51479bde0ee36732f2d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
