+++ Assets/ScriptableObjects/ItemData/item_atk5.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4bdb36b2e6c2c1d408548cd198f87044
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
