+++ Assets/ScriptableObjects/ItemData/item_firerate.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2279cffc04ed1214ab4750319117313c, type: 3}
+  m_Name: item_firerate
+  m_EditorClassIdentifier: 
+  displayName: "\uC5F0\uC0AC\uC18D\uB3C4x1.2"
+  description: "\uC5F0\uC0AC\uC18D\uB3C4\uAC00 1.2\uBC30\uB85C \uC99D\uAC00"
+  Stype: 1
+  calculate: 1
+  value: 1.2
+  cost: 5000
