+++ Assets/ScriptableObjects/ItemData/item_firerate.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48dfa09a477657147b0a98bf8e73c7a3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
