+++ Assets/ScriptableObjects/ItemData/item_hp10.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 37d43ea06ac48b54ebef6e990dcfbd1c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
