+++ Assets/ScriptableObjects/ItemData/item_hp5.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2279cffc04ed1214ab4750319117313c, type: 3}
+  m_Name: item_hp5
+  m_EditorClassIdentifier: 
+  displayName: "\uCCB4\uB825 +5"
+  description: "\uCCB4\uB825\uC774 5\uB9CC\uD07C \uC99D\uAC00\uD55C\uB2E4"
+  Stype: 2
+  calculate: 0
+  value: 5
+  cost: 1000
