+++ Assets/ScriptableObjects/ItemData/item_hp5.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: e6f67d1a2a69a454c979d078ee75f786
-folderAsset: yes
-DefaultImporter:
+guid: f46b64a7a727a6b40bdd71f7127bfd4c
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
