+++ Assets/Scripts/KCW/StoreDataManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5c211402d638c744684109db7be79803
+guid: 3c7cd1b9e8852c749a0c1a55f48ea799
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
